Add files via upload
This SQL query file is part of my SQL portfolio and complements the 'instagram_database_clone_project.sql' file uploaded in this repository. This file provides the structure and relationships for the database used in the queries. Please refer to that file to get a comprehensive understanding of the database design.